Interactively plotting 3D attainment plots

Interactively plotting 3D attainment plots

Usage

interactive3dplot(
  x,
  percentiles = c(0, 50, 100),
  col = c("black", "white"),
  type = "point",
  psize = 3,
  xlabel = "X",
  ylabel = "Y",
  zlabel = "Z"
)

Arguments

x

Either a matrix of data values, or a data frame, or a list of data frames of exactly three columns.

percentiles

(numeric()) - Vector indicating which percentile should be plot. The default is to plot only the median attainment curve. If type='cube' then only the first percentile will be used.

col

A vector of colours used to plot.

type

(character(1)) - string giving the type of plot desired. The following values are possible, 'scatter' and 'cube'.

psize

Size of plotted points.

xlabel, ylabel, zlabel

Axes labels.

Value

A 3D plot of the attainment points and surfaces

Examples

options("warn"=-1)
library(interactiveeafplots)

extdata_dir <- system.file(package="interactiveeafplots", "extdata")
x <- moocore::read_datasets(file.path(extdata_dir, "spherical-250-10-3d.txt"))
# Select only the last 3 datasets
x <- x[x[, 4] >= 8, ]

interactive3dplot(x, c(0,50,100), col=c("red","orange","yellow"),
                  type="scatter", psize=3, xlabel="Objective 1",
                  ylabel="Objective 2", zlabel="Objective 3")

# Will take a long time to render.
# Opens in separate window, so run by yourself.
# interactive3dplot(x, c(0,50,100), col=c("red","yellow"), type="cube",
#                   psize=3, xlabel="Objective 1", ylabel="Objective 2",
#                   zlabel="Objective 3")