Interactively plotting the EAF

Interactively plotting the EAF

Usage

interactiveeafplot(
  x,
  percentiles = c(0, 50, 100),
  maximise = FALSE,
  col = c("white", "black"),
  type = "point",
  lty = "solid",
  psize = 2,
  pshape = 16,
  legend.pos = "topright",
  xaxis.side = "below",
  yaxis.side = "left",
  axes = TRUE,
  xlabel = NULL,
  ylabel = NULL,
  sci.notation = FALSE,
  plot = "plotly"
)

Arguments

x

Either a matrix of data values, or a data frame, or a list of data frames of exactly three columns.

percentiles

(numeric()) - Vector indicating which percentile should be plot. The default is to plot only the median attainment curve.

maximise

(logical() | logical(1)) - Whether the objectives must be maximised instead of minimised. Either a single logical value that applies to all objectives or a vector of logical values, with one value per objective.

col

A vector of colours used to plot.

type

(character(1)) - string giving the type of plot desired. The following values are possible, 'point' and 'area'.

lty

Line type. See plot.default().

psize

Size of plotted points

pshape

Shape of plotted point. See plot.default().

legend.pos

The position of the legend, valid values are 'topright', 'top', 'topleft', 'right', 'left', 'bottom', 'bottomright', 'bottomleft' and 'center'. A value of 'none' hides the legend.

xaxis.side

On which side that xaxis is drawn. Valid values are 'below' and 'above'.

yaxis.side

On which side that yaxis is drawn. Valid values are 'left' and 'right'.

axes

A logical value indicating whether both axes should be drawn on the plot.

xlabel, ylabel

Axes labels.

sci.notation

Generate scientific labels

plot

Interactive or none interactive plot. Valid values are 'plotly' and 'ggplot'.

Value

A plot of the attainment surfaces.

Examples

options("warn"=-1)
library(interactiveeafplots)

data(gcp2x2,package='interactiveeafplots')
tabucol <- subset(gcp2x2, alg != "TSinN1")
tabucol$alg <- tabucol$alg[drop=TRUE]
mydata <- dplyr::filter(tabucol, inst == "DSJC500.5")
mydata <- mydata[c("time","best","run")]
interactiveeafplot(x=mydata, percentiles=c(0,50,100), col=c("yellow","red"),
                   maximise=FALSE, type="area", lty="longdash",
                   psize=3, pshape=10, legend.pos="topright",
                   xaxis.side="top", yaxis.side="left", axes=TRUE,
                   sci.notation=FALSE, xlabel="MIN X", ylabel="MIN Y",
                   plot="plotly")

data(SPEA2minstoptimeRichmond,package='interactiveeafplots')
SPEA2minstoptimeRichmond[,2] <- SPEA2minstoptimeRichmond[,2] / 60
interactiveeafplot(SPEA2minstoptimeRichmond, percentiles = c(0,50,100),
                   col=c("yellow","red"),maximise=c(TRUE,FALSE), type="point",
                   lty="longdash", psize=3, pshape=10, legend.pos="topleft",
                   xaxis.side="bottom", yaxis.side="left", axes=TRUE,
                   sci.notation=FALSE, xlabel = "C[E]",
                   ylabel = "Minimum idle time (minutes)",
                   plot = "plotly")